requirements.txt
discord.py[speed]<3.0.0
aiohttp[speedups]
pymongo
objgraph
validators
colour
numpy
repoze.timeago
jsonpickle<3.0.4
emoji
Pillow
beautifulsoup4
sentry_sdk
pydealer
ssdeep
aiohttp_socks
cython
psutil
setuptools
py-cpuinfo
black
audioop-lts
mongoengine
